requirements.txt
python-dotenv
rq
telebot
pytz
Flask
python-telegram-bot
Flask-Cors==3.0.10
gunicorn==20.1.0
Jinja2==3.1.2
Flask-HTTPAuth
bcrypt
ccxt
networkx
Cython
lxml
requests
telegram